+++ src/main/scala/edu/ie3/powerFactory2psdm/converter/types/LineTypeConverter.scala
@@ -6,7 +6,7 @@
 
 package edu.ie3.powerFactory2psdm.converter.types
 import edu.ie3.datamodel.models.input.connector.`type`.LineTypeInput
-import edu.ie3.powerFactory2psdm.model.powerfactory.types.LineType
+import edu.ie3.powerFactory2psdm.model.entity.types.LineType
 import tech.units.indriya.quantity.Quantities
 import edu.ie3.util.quantities.PowerSystemUnits.{
   KILOVOLT,
