+++ src/main/scala/edu/ie3/powerFactory2psdm/io/PfGridParser.scala
@@ -7,10 +7,11 @@
 package edu.ie3.powerFactory2psdm.io
 
 import com.typesafe.scalalogging.LazyLogging
-import edu.ie3.powerFactory2psdm.model.powerfactory.RawPfGridModel
+import edu.ie3.powerFactory2psdm.model.RawPfGridModel
 import io.circe.parser.decode
 import io.circe.generic.auto._
 import io.circe.parser._
+
 import scala.io.Source
 import java.io.File
