+++ src/main/scala/edu/ie3/powerFactory2psdm/model/PreprocessedPfGridModel.scala
@@ -4,28 +4,14 @@
  * Research group Distribution grid planning and operation
  */
 
-package edu.ie3.powerFactory2psdm.model.powerfactory
+package edu.ie3.powerFactory2psdm.model
 
 import com.typesafe.scalalogging.LazyLogging
-import edu.ie3.powerFactory2psdm.exception.pf.{
-  ConversionException,
-  GridConfigurationException,
-  MissingParameterException
-}
-import edu.ie3.powerFactory2psdm.model.powerfactory.RawPfGridModel.{
-  LineTypes,
-  Lines,
-  Nodes,
-  ProjectSettings,
-  Switches,
-  TrafoTypes2w,
-  Trafos2w
-}
-import edu.ie3.powerFactory2psdm.model.powerfactory.types.{
-  LineType,
-  TransformerType2W
-}
-import edu.ie3.powerFactory2psdm.util.ConversionPrefixes
+import edu.ie3.powerFactory2psdm.exception.pf.{ConversionException, GridConfigurationException, MissingParameterException}
+import edu.ie3.powerFactory2psdm.model.RawPfGridModel.{LineTypes, Lines, Nodes, ProjectSettings, Switches, TrafoTypes2w, Trafos2w}
+import edu.ie3.powerFactory2psdm.model.entity.{Line, Node, Switch}
+import edu.ie3.powerFactory2psdm.model.entity.types.{LineType, TransformerType2W}
+import edu.ie3.powerFactory2psdm.model.setting.ConversionPrefixes
 
 /**
   * Representation of the grid which is to be converted to a PSDM [[edu.ie3.datamodel.models.input.container.GridContainer]].
