+++ src/main/scala/edu/ie3/powerFactory2psdm/model/entity/Edge.scala
@@ -4,7 +4,7 @@
  * Research group Distribution grid planning and operation
  */
 
-package edu.ie3.powerFactory2psdm.model.powerfactory
+package edu.ie3.powerFactory2psdm.model.entity
 
 /**
   * Denotes an element that connects two nodes within a Subnet
