+++ src/main/scala/edu/ie3/powerFactory2psdm/model/entity/Subnet.scala
@@ -4,16 +4,15 @@
  * Research group Distribution grid planning and operation
  */
 
-package edu.ie3.powerFactory2psdm.model
+package edu.ie3.powerFactory2psdm.model.entity
 
 import edu.ie3.datamodel.models.voltagelevels.VoltageLevel
-import edu.ie3.powerFactory2psdm.model.powerfactory.Node
 
 /**
   * Data type that wraps necessary information for the different subnets
   *
-  * @param id of the subnet
-  * @param nodeIds UUIDS of the nodes inside the subnet
+  * @param id      of the subnet
+  * @param nodes   UUIDS of the nodes inside the subnet
   * @param voltLvl voltage level that the nodes live in
   */
 final case class Subnet(id: Int, nodes: Set[Node], voltLvl: VoltageLevel)
