+++ src/test/scala/edu/ie3/powerFactory2psdm/common/ConverterTestData.scala
@@ -7,40 +7,28 @@
 package edu.ie3.powerFactory2psdm.common
 
 import com.typesafe.scalalogging.LazyLogging
-import edu.ie3.datamodel.models.input.connector.`type`.{
-  LineTypeInput,
-  Transformer2WTypeInput
-}
+import edu.ie3.datamodel.models.input.connector.`type`.{LineTypeInput, Transformer2WTypeInput}
 import edu.ie3.datamodel.models.{OperationTime, StandardUnits, UniqueEntity}
 import edu.ie3.datamodel.models.input.{NodeInput, OperatorInput}
 import edu.ie3.datamodel.models.voltagelevels.GermanVoltageLevelUtils.LV
 import edu.ie3.powerFactory2psdm.config.ConversionConfig
+
 import java.io.File
 import edu.ie3.powerFactory2psdm.exception.io.GridParsingException
 import edu.ie3.powerFactory2psdm.exception.pf.TestException
 import edu.ie3.powerFactory2psdm.io.PfGridParser
-import edu.ie3.powerFactory2psdm.model.Subnet
-import edu.ie3.powerFactory2psdm.model.powerfactory.types.{
-  LineType,
-  TransformerType2W
-}
-import edu.ie3.powerFactory2psdm.model.powerfactory.{
-  ConnectedElement,
-  EntityModel,
-  PreprocessedPfGridModel,
-  Node
-}
+import edu.ie3.powerFactory2psdm.model.entity.{ConnectedElement, EntityModel, Node, Subnet}
+import edu.ie3.powerFactory2psdm.model.entity.types.{LineType, TransformerType2W}
+import edu.ie3.powerFactory2psdm.model.PreprocessedPfGridModel
+import edu.ie3.powerFactory2psdm.model.types.TransformerType2W
 import edu.ie3.util.quantities.PowerSystemUnits.PU
 import org.locationtech.jts.geom.{Coordinate, GeometryFactory}
 import pureconfig.ConfigSource
 import tech.units.indriya.quantity.Quantities
 import tech.units.indriya.unit.Units.{OHM, PERCENT, SIEMENS}
-import edu.ie3.util.quantities.PowerSystemUnits.{
-  DEGREE_GEOM,
-  KILOVOLT,
-  VOLTAMPERE
-}
+import edu.ie3.util.quantities.PowerSystemUnits.{DEGREE_GEOM, KILOVOLT, VOLTAMPERE}
 import pureconfig.generic.auto._
+
 import java.util.UUID
 import javax.measure.MetricPrefix
