+++ src/test/scala/edu/ie3/powerFactory2psdm/model/PreprocessedPfGridModelSpec.scala
@@ -4,14 +4,11 @@
  * Research group Distribution grid planning and operation
  */
 
-package edu.ie3.powerFactory2psdm.model.powerfactory
+package edu.ie3.powerFactory2psdm.model
 
-import edu.ie3.powerFactory2psdm.exception.pf.{
-  ConversionException,
-  MissingParameterException
-}
+import edu.ie3.powerFactory2psdm.exception.pf.{ConversionException, MissingParameterException}
 import edu.ie3.powerFactory2psdm.model.powerfactory.RawPfGridModel.ProjectSettings
-import edu.ie3.powerFactory2psdm.util.ConversionPrefixes
+import edu.ie3.powerFactory2psdm.model.setting.ConversionPrefixes
 import org.scalatest.PrivateMethodTester
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
