+++ src/test/scala/edu/ie3/powerFactory2psdm/model/types/TransformerType2WSpec.scala
@@ -4,11 +4,11 @@
  * Research group Distribution grid planning and operation
  */
 
-package edu.ie3.powerFactory2psdm.model.powerfactory
+package edu.ie3.powerFactory2psdm.model.types
 
 import edu.ie3.powerFactory2psdm.exception.pf.MissingParameterException
-import edu.ie3.powerFactory2psdm.model.powerfactory.RawPfGridModel.TrafoTypes2w
-import edu.ie3.powerFactory2psdm.model.powerfactory.types.TransformerType2W
+import edu.ie3.powerFactory2psdm.model.RawPfGridModel.TrafoTypes2w
+import edu.ie3.powerFactory2psdm.model.entity.types.TransformerType2W
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
